refactor: App 컴포넌트 리팩토링 및 상수화
