+++ src/App.js
@@ -1,17 +1,24 @@
-import GlobalStyles from './global.styles';
 import React, { useState } from 'react';
-import Card from './common/Card';
-import Modal from './common/Modal';
-import Nav from './components/Nav';
-
 import { AppWrapper } from './App.styles.js';
+import GlobalStyles from './global.styles';
+
 import NewCardForm from './components/NewCardForm';
+import CardAdditionComplete from './components/CardAdditionComplete';
+import Nav from './components/mixin/Nav';
 import CardColor from './components/ModalContents/CardColor';
 import CVCHelp from './components/ModalContents/CVCHelp';
-import CardAdditionComplete from './components/CardAdditionComplete';
+
+import Card from './common/Card';
+import Modal from './common/Modal';
+
+import { MODAL, PAGE } from './constants/constant';
 
 function App() {
-  const [page, setPage] = useState('addCard');
+  const [page, setPage] = useState(PAGE.ADD_CARD);
+  const [openModalContent, setOpenModalContent] = useState({
+    isModalOpen: true,
+    modalContent: MODAL.CARD_COLOR,
+  });
   const [myCards, setMyCards] = useState([]);
   const [newCardInfo, setNewCardInfo] = useState({
     cardName: 'DEFAULT',
@@ -33,6 +40,19 @@ function App() {
       second: '',
     },
   });
+
+  const addNewCard = () => {
+    setMyCards([...myCards, newCardInfo]);
+    resetNewCardInfo();
+  };
+
+  const handleCardColor = (name) => {
+    setNewCardInfo({
+      ...newCardInfo,
+      cardName: name,
+    });
+    handleModalClose();
+  };
 
   const resetNewCardInfo = () => {
     setNewCardInfo({
@@ -57,33 +77,10 @@ function App() {
     });
   };
 
-  const addNewCard = () => {
-    setMyCards([...myCards, newCardInfo]);
-    resetNewCardInfo();
-  };
-
-  const handleCardColor = (name) => {
-    setNewCardInfo({
-      ...newCardInfo,
-      cardName: name,
-    });
-    handleModalClose();
-  };
-
-  const modalContentsObject = {
-    cardColor: <CardColor handleCardColor={handleCardColor} />,
-    cvcHelp: <CVCHelp />,
-  };
-
-  const [openModalContent, setOpenModalContent] = useState({
-    isModalOpen: true,
-    modalContent: modalContentsObject.cardColor,
-  });
-
-  const handleModalOpen = (content) => {
+  const handleModalOpen = (modalContent) => {
     setOpenModalContent({
       isModalOpen: true,
-      modalContent: modalContentsObject[content],
+      modalContent,
     });
   };
 
@@ -98,7 +95,7 @@ function App() {
     <>
       <GlobalStyles />
       <AppWrapper>
-        {page === 'addCard' && (
+        {page === PAGE.ADD_CARD && (
           <>
             <Nav />
             <div className='card-wrapper'>
@@ -112,7 +109,7 @@ function App() {
             />
           </>
         )}
-        {page === 'cardComplete' && (
+        {page === PAGE.CARD_COMPLETE && (
           <CardAdditionComplete
             newCardInfo={newCardInfo}
             setNewCardInfo={setNewCardInfo}
@@ -122,7 +119,10 @@ function App() {
 
         {openModalContent.isModalOpen && (
           <Modal handleModalClose={handleModalClose}>
-            {openModalContent.modalContent}
+            {openModalContent.modalContent === MODAL.CARD_COLOR && (
+              <CardColor handleCardColor={handleCardColor} />
+            )}
+            {openModalContent.modalContent === MODAL.CVC_HELP && <CVCHelp />}
           </Modal>
         )}
       </AppWrapper>
