+++ src/components/mixin/Nav/index.styles.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { FONT_SIZE } from '../../constants/constant';
+import { FONT_SIZE } from '../../../constants/constant';
 
 const NavWrapper = styled.div`
   width: 100%;
